Encrypt messages between nodes
Both sides of the chat now generate secrets based on their private key
and the other side's public key. They encrypt each message with a nonce
before sending it, and use the nonce in the decryption.

This _does_ mean that the responder has to specify the starter's name
now so that they can select the proper public key for encryption.