Document a way to get Veilid from source on Windows
This provides one method for getting the Veilid source code on
Windows, and launching a node. It is mainly intended to confirm that
https://gitlab.com/veilid/cursive-flexi-logger-view/-/merge_requests/1
addresses a build problem.

It provides a "development" environment in the sense that one could,
in theory, edit the .rs files in Notepad. It does not discuss whether
there is any value in seeing the debug messages in a cmd.exe window,
and does not document other deployment options that may be possible,
such as a Windows service. The information could be revised later to
be more maintainable, e.g., by avoiding the hardcoded protoc and capnp
version numbers and removing the x86_64 assumption.