Require Cap’n Proto 1.0.1 on Windows
This changes the Windows instructions to specify
capnproto-c++-win32-1.0.1. Previously, users may have been trying to
use Veilid with 0.10.4, especially in the common case of using the
winget package manager.

Also, the suggested Protocol Buffers version is updated to 24.3, and a
note is added about rustup behavior on Windows machines that have
never previously been used for development.