Added --delete option to rsync push to repo server
The rsync push successfully put the new repo files on the repo
server, but also left the original files in place which caused
old packages and repo errata to build up over time. Adding the
--delete switch results removal of files on the repo server
that were deleted as part of the the distribution scripts
"Clean and reset the workspaces" section.